version test mais q mediocre

+++ libavcodec/bc6enc.h
@@ -21,9 +21,9 @@ typedef enum BC6EncInputFormat {
 } BC6EncInputFormat;
 
 void ff_bc6enc_init(BC6EncContext *c, int is_signed, uint16_t mode_mask,
-                    float exposure, float quality, int use_pattern);
+                    float exposure, int quality, int use_pattern);
 void ff_bc6enc_init_input(BC6EncContext *c, int is_signed, uint16_t mode_mask,
-                          float exposure, float quality, int use_pattern,
+                          float exposure, int quality, int use_pattern,
                           BC6EncInputFormat input_fmt, int max_try);
 
 #endif /* AVCODEC_BC6ENC_H */
