+++ libavcodec/hap.h
@@ -70,7 +70,7 @@ typedef struct HapContext {
     int opt_chunk_count; /* User-requested chunk count (encoder only) */
     int opt_compressor; /* User-requested compressor (encoder only) */
     int opt_bc7_quality; /* BC7 encoder quality level (encoder only) */
-    float opt_bc6_quality; /* BC6 encoder quality level (encoder only) */
+    int opt_bc6_quality; /* BC6 encoder quality level (encoder only) */
 
     int chunk_count;
     HapChunk *chunks;
